requirements.txt
requests
google-genai
openai
